refactor(profiling): add explicit casts to avoid narrowing conversions
Use explicit static_cast where implicit narrowing conversions occur:
- danger.cc: Cast page calculations to safe_memcpy_return_t
- mirrors.cc: Cast table_size calculation to ssize_t
- greenlets.cc: Cast count to int before adding
- stack_renderer.cpp: Cast microsecond_t to int64_t for push_* calls
- vm.cc: Cast size_t to off_t for ftruncate, riovcnt to int for pwritev
- uploader.cpp: Cast len to std::streamsize for write()
- frame.cc: Cast offsetof result to int
- sampler.cpp: Cast uint64_t to double before subtraction

Fixes clang-tidy bugprone-narrowing-conversions warnings.

+++ ddtrace/internal/datadog/profiling/stack/src/echion/frame.cc
@@ -269,7 +269,7 @@ Frame::read(EchionSampler& echion, PyObject* frame_addr, PyObject** prev_addr)
       (static_cast<int>(
         (frame_addr->instr_ptr - 1 -
          reinterpret_cast<_Py_CODEUNIT*>((reinterpret_cast<PyCodeObject*>(frame_addr->f_executable)))))) -
-      offsetof(PyCodeObject, co_code_adaptive) / sizeof(_Py_CODEUNIT);
+      static_cast<int>(offsetof(PyCodeObject, co_code_adaptive) / sizeof(_Py_CODEUNIT));
     auto maybe_frame = Frame::get(echion, reinterpret_cast<PyCodeObject*>(frame_addr->f_executable), lasti);
     if (!maybe_frame) {
         return ErrorKind::FrameError;
