+++ ddtrace/internal/datadog/profiling/stack/src/stack_renderer.cpp
@@ -55,7 +55,7 @@ StackRenderer::render_thread_begin(PyThreadState* tstate,
 
     // Finalize the thread information we have
     sample->push_threadinfo(static_cast<int64_t>(thread_id), static_cast<int64_t>(native_id), name);
-    sample->push_walltime(thread_state.wall_time_ns, 1);
+    sample->push_walltime(static_cast<int64_t>(thread_state.wall_time_ns), 1);
 
     const std::optional<Span> active_span = ThreadSpanLinks::get_instance().get_active_span_from_thread_id(thread_id);
     if (active_span) {
@@ -87,11 +87,11 @@ StackRenderer::render_task_begin(const std::string& task_name, bool on_cpu)
         // Add the thread context into the sample
         sample->push_threadinfo(
           static_cast<int64_t>(thread_state.id), static_cast<int64_t>(thread_state.native_id), thread_state.name);
-        sample->push_walltime(thread_state.wall_time_ns, 1);
+        sample->push_walltime(static_cast<int64_t>(thread_state.wall_time_ns), 1);
 
         if (on_cpu) {
             // initialized to 0, so possibly a no-op
-            sample->push_cputime(thread_state.cpu_time_ns, 1);
+            sample->push_cputime(static_cast<int64_t>(thread_state.cpu_time_ns), 1);
         }
 
         sample->push_monotonic_ns(thread_state.now_time_ns);
@@ -225,7 +225,7 @@ StackRenderer::render_cpu_time(uint64_t cpu_time_us)
     // TODO - it's absolutely false that thread-level CPU time is task time.  This needs to be normalized
     // to the task level, but for now just keep it because this is how the v1 sampler works
     thread_state.cpu_time_ns = 1000LL * cpu_time_us;
-    sample->push_cputime(thread_state.cpu_time_ns, 1);
+    sample->push_cputime(static_cast<int64_t>(thread_state.cpu_time_ns), 1);
 }
 
 void
